+++ AUTHORS
@@ -28,3 +28,4 @@ Here is a list of passed and present much-appreciated contributors:
     Tommy Anthony
     Tsuyoshi Hombashi
     Tushar Makkar
+    Peyman Salehi
