+++ src/tablib/formats/_df.py
@@ -11,7 +11,6 @@ except ImportError:
 
 import tablib
 
-from tablib.compat import unicode
 
 title = 'df'
 extensions = ('df', )
