+++ src/tablib/formats/_yaml.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 """ Tablib - YAML Support.
 """
