+++ src/tablib/packages/dbfpy/dbf.py
@@ -64,10 +64,10 @@ __all__ = ["Dbf"]
 
 from . import header
 from . import record
-from utils import INVALID_VALUE
+from .utils import INVALID_VALUE
 
 
-class Dbf(object):
+class Dbf:
     """DBF accessor.
 
     FIXME:
@@ -114,13 +114,13 @@ class Dbf(object):
                 ``INVALID_VALUE`` instead of raising conversion error.
 
         """
-        if isinstance(f, basestring):
+        if isinstance(f, str):
             # a filename
             self.name = f
             if new:
                 # new table (table file must be
                 # created or opened and truncated)
-                self.stream = file(f, "w+b")
+                self.stream = open(f, "w+b")
             else:
                 # table file must exist
                 self.stream = file(f, ("r+b", "rb")[bool(readOnly)])
@@ -177,7 +177,7 @@ class Dbf(object):
             Return value is numeric object maning valid index.
 
         """
-        if not isinstance(index, (int, long)):
+        if not isinstance(index, int):
             raise TypeError("Index must be a numeric object")
         if index < 0:
             # index from the right side
@@ -204,7 +204,8 @@ class Dbf(object):
     def indexOfFieldName(self, name):
         """Index of field named ``name``."""
         # FIXME: move this to header class
-        return self.header.fields.index(name)
+        names = [f.name for f in self.header.fields]
+        return names.index(name.upper())
 
     def newRecord(self):
         """Return new record, which belong to this table."""
@@ -294,4 +295,4 @@ if __name__ == '__main__':
     demo_create(_name)
     demo_read(_name)
 
-    # vim: set et sw=4 sts=4 :
+# vim: set et sw=4 sts=4 :
