+++ src/tablib/packages/dbfpy/record.py
@@ -16,11 +16,11 @@ __date__ = "$Date: 2007/02/11 09:05:49 $"[7:-2]
 
 __all__ = ["DbfRecord"]
 
-from itertools import izip
+import sys
 
-import utils
+from . import utils
 
-class DbfRecord(object):
+class DbfRecord:
     """DBF record.
 
     Instances of this class shouldn't be created manually,
@@ -170,13 +170,14 @@ class DbfRecord(object):
         """
         self._validateIndex(False)
         self.dbf.stream.seek(self.position)
-        self.dbf.stream.write(self.toString())
+        self.dbf.stream.write(bytes(self.toString(),
+            sys.getfilesystemencoding()))
         # FIXME: may be move this write somewhere else?
         # why we should check this condition for each record?
         if self.index == len(self.dbf):
             # this is the last record,
             # we should write SUB (ASCII 26)
-            self.dbf.stream.write("\x1A")
+            self.dbf.stream.write(b"\x1A")
 
     ## utility methods
 
@@ -218,9 +219,12 @@ class DbfRecord(object):
 
     def toString(self):
         """Return string packed record values."""
+#        for (_def, _dat) in zip(self.dbf.header.fields, self.fieldData):
+#
+
         return "".join([" *"[self.deleted]] + [
-            _def.encodeValue(_dat)
-            for (_def, _dat) in izip(self.dbf.header.fields, self.fieldData)
+           _def.encodeValue(_dat)
+            for (_def, _dat) in zip(self.dbf.header.fields, self.fieldData)
         ])
 
     def asList(self):
@@ -241,11 +245,11 @@ class DbfRecord(object):
             real values stored in this object.
 
         """
-        return dict([_i for _i in izip(self.dbf.fieldNames, self.fieldData)])
+        return dict([_i for _i in zip(self.dbf.fieldNames, self.fieldData)])
 
     def __getitem__(self, key):
         """Return value by field name or field index."""
-        if isinstance(key, (long, int)):
+        if isinstance(key, int):
             # integer index of the field
             return self.fieldData[key]
         # assuming string field name
@@ -253,7 +257,7 @@ class DbfRecord(object):
 
     def __setitem__(self, key, value):
         """Set field value by integer index of the field or string name."""
-        if isinstance(key, (int, long)):
+        if isinstance(key, int):
             # integer index of the field
             return self.fieldData[key]
         # assuming string field name
