+++ src/tablib/packages/dbfpy/utils.py
@@ -25,7 +25,7 @@ def unzfill(str):
 
     """
     try:
-        return str[:str.index('\0')]
+        return str[:str.index(b'\0')]
     except ValueError:
         return str
 
@@ -60,10 +60,10 @@ def getDate(date=None):
         return date
     if isinstance(date, datetime.datetime):
         return date.date()
-    if isinstance(date, (int, long, float)):
+    if isinstance(date, (int, float)):
         # date is a timestamp
         return datetime.date.fromtimestamp(date)
-    if isinstance(date, basestring):
+    if isinstance(date, str):
         date = date.replace(" ", "0")
         if len(date) == 6:
             # yymmdd
@@ -107,10 +107,10 @@ def getDateTime(value=None):
         return value
     if isinstance(value, datetime.date):
         return datetime.datetime.fromordinal(value.toordinal())
-    if isinstance(value, (int, long, float)):
+    if isinstance(value, (int, float)):
         # value is a timestamp
         return datetime.datetime.fromtimestamp(value)
-    if isinstance(value, basestring):
+    if isinstance(value, str):
         raise NotImplementedError("Strings aren't currently implemented")
     if hasattr(value, "__getitem__"):
         # a sequence (assuming date/time tuple)
@@ -145,7 +145,7 @@ class _InvalidValue(object):
     def __ne__(self, other):
         return not (other is self)
 
-    def __nonzero__(self):
+    def __bool__(self):
         return False
 
     def __int__(self):
@@ -159,7 +159,7 @@ class _InvalidValue(object):
         return ""
 
     def __unicode__(self):
-        return u""
+        return ""
 
     def __repr__(self):
         return "<INVALID>"
