+++ tests/requirements.txt
@@ -1,6 +1,5 @@
 pytest
 pytest-cov
-backports.csv; python_version < '3.0'
 MarkupPy
 odfpy
 openpyxl>=2.4.0
