+++ tox.ini
@@ -2,12 +2,11 @@
 usedevelop = true
 minversion = 2.4
 envlist =
-    py{27,35,36,37,38}-tests,
+    py{35,36,37,38}-tests,
     py37-{docs,lint}
 
 [testenv]
 basepython =
-    py27: python2.7
     py35: python3.5
     py36: python3.6
     py37: python3.7
@@ -38,3 +37,4 @@ commands =
 exclude =
     .tox
 ignore=E501,E127,E128,E124
+
